[GL,FileIO] Move Legacy folder, {read,write}GeometryToFile from GeoLib/IO to Applications/FileIO.
