+++ Applications/FileIO/CMakeLists.txt
@@ -1,5 +1,6 @@
 GET_SOURCE_FILES(SOURCES)
 APPEND_SOURCE_FILES(SOURCES Gmsh)
+APPEND_SOURCE_FILES(SOURCES Legacy)
 
 if(NOT Shapelib_FOUND)
     list(REMOVE_ITEM SOURCES SHPInterface.h SHPInterface.cpp)
