+++ Applications/Utils/MeshEdit/CreateBoundaryConditionsAlongPolylines.cpp
@@ -17,6 +17,8 @@
 #include <tclap/CmdLine.h>
 
 #include "Applications/ApplicationsLib/LogogSetup.h"
+#include "Applications/FileIO/readGeometryFromFile.h"
+#include "Applications/FileIO/writeGeometryToFile.h"
 
 #include "BaseLib/FileTools.h"
 
@@ -25,8 +27,6 @@
 
 #include "GeoLib/GEOObjects.h"
 #include "GeoLib/Point.h"
-#include "GeoLib/IO/readGeometryFromFile.h"
-#include "GeoLib/IO/writeGeometryToFile.h"
 
 #include "MeshLib/Mesh.h"
 #include "MeshLib/Node.h"
@@ -90,9 +90,9 @@ void writeBCsAndGeometry(GeoLib::GEOObjects& geometry_sets,
 {
     if (write_gml) {
         INFO("write points to \"%s.gml\".", geo_name.c_str());
-        GeoLib::IO::writeGeometryToFile(geo_name, geometry_sets, out_fname+".gml");
+        FileIO::writeGeometryToFile(geo_name, geometry_sets, out_fname+".gml");
     }
-    GeoLib::IO::writeGeometryToFile(geo_name, geometry_sets, out_fname+".gli");
+    FileIO::writeGeometryToFile(geo_name, geometry_sets, out_fname+".gli");
 
     bool liquid_flow(false);
     if (bc_type == "LIQUID_FLOW")
@@ -177,7 +177,7 @@ int main (int argc, char* argv[])
 
     // *** read geometry
     GeoLib::GEOObjects geometries;
-    GeoLib::IO::readGeometryFromFile(geometry_fname.getValue(), geometries);
+    FileIO::readGeometryFromFile(geometry_fname.getValue(), geometries);
 
     std::string geo_name;
     {
