+++ GeoLib/CMakeLists.txt
@@ -2,7 +2,6 @@
 GET_SOURCE_FILES(SOURCES)
 
 APPEND_SOURCE_FILES(SOURCES IO)
-APPEND_SOURCE_FILES(SOURCES IO/Legacy)
 
 APPEND_SOURCE_FILES(SOURCES IO/XmlIO/Rapid)
 APPEND_SOURCE_FILES(SOURCES IO/XmlIO/Boost)
