+++ Tests/GeoLib/IO/TestGLIReader.cpp
@@ -16,7 +16,7 @@
 #include "gtest/gtest.h"
 
 #include "BaseLib/BuildInfo.h"
-#include "GeoLib/IO/Legacy/OGSIOVer4.h"
+#include "Applications/FileIO/Legacy/OGSIOVer4.h"
 #include "GeoLib/GEOObjects.h"
 
 class OGSIOVer4InterfaceTest : public ::testing::Test
@@ -55,7 +55,8 @@ TEST_F(OGSIOVer4InterfaceTest, SimpleTIN)
     GeoLib::GEOObjects geometries;
     std::vector<std::string> errors;
     std::string geometry_name("TestGeometry");
-    GeoLib::IO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name, errors);
+    FileIO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name,
+                                  errors);
 
     std::vector<GeoLib::Surface*> const*
         sfcs(geometries.getSurfaceVec(geometry_name));
@@ -78,7 +79,8 @@ TEST_F(OGSIOVer4InterfaceTest, StillCorrectTINWihtAdditionalValueAtEndOfLine)
     GeoLib::GEOObjects geometries;
     std::vector<std::string> errors;
     std::string geometry_name("TestGeometry");
-    GeoLib::IO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name, errors);
+    FileIO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name,
+                                  errors);
 
     std::vector<GeoLib::Surface*> const*
         sfcs(geometries.getSurfaceVec(geometry_name));
@@ -100,7 +102,8 @@ TEST_F(OGSIOVer4InterfaceTest, InvalidTIN_ZeroAreaTri)
     GeoLib::GEOObjects geometries;
     std::vector<std::string> errors;
     std::string geometry_name("TestGeometry");
-    GeoLib::IO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name, errors);
+    FileIO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name,
+                                  errors);
 
     std::vector<GeoLib::Surface*> const*
         sfcs(geometries.getSurfaceVec(geometry_name));
@@ -123,7 +126,8 @@ TEST_F(OGSIOVer4InterfaceTest, InvalidTIN_LineDoesNotStartWithID)
     GeoLib::GEOObjects geometries;
     std::vector<std::string> errors;
     std::string geometry_name("TestGeometry");
-    GeoLib::IO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name, errors);
+    FileIO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name,
+                                  errors);
 
     std::vector<GeoLib::Surface*> const*
         sfcs(geometries.getSurfaceVec(geometry_name));
@@ -145,7 +149,8 @@ TEST_F(OGSIOVer4InterfaceTest, InvalidTIN_PointIsMissing)
     GeoLib::GEOObjects geometries;
     std::vector<std::string> errors;
     std::string geometry_name("TestGeometry");
-    GeoLib::IO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name, errors);
+    FileIO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name,
+                                  errors);
 
     std::vector<GeoLib::Surface*> const*
         sfcs(geometries.getSurfaceVec(geometry_name));
@@ -166,7 +171,8 @@ TEST_F(OGSIOVer4InterfaceTest, InvalidTIN_CoordOfPointIsMissing)
     GeoLib::GEOObjects geometries;
     std::vector<std::string> errors;
     std::string geometry_name("TestGeometry");
-    GeoLib::IO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name, errors);
+    FileIO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name,
+                                  errors);
 
     std::vector<GeoLib::Surface*> const*
         sfcs(geometries.getSurfaceVec(geometry_name));
@@ -187,7 +193,8 @@ TEST_F(OGSIOVer4InterfaceTest, SimpleTIN_AdditionalEmptyLinesAtEnd)
     GeoLib::GEOObjects geometries;
     std::vector<std::string> errors;
     std::string geometry_name("TestGeometry");
-    GeoLib::IO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name, errors);
+    FileIO::Legacy::readGLIFileV4(_gli_fname, geometries, geometry_name,
+                                  errors);
 
     std::vector<GeoLib::Surface*> const*
         sfcs(geometries.getSurfaceVec(geometry_name));
